# Fix perf regression in ShuffleRows